fix(pipelines): keep host policy client-safe (#1356)
